refactor: extract ymax-tool-lib
