+++ packages/portfolio-deploy/package.json
@@ -48,7 +48,8 @@
     "@endo/pass-style": "^1.6.3",
     "@endo/patterns": "^1.7.0",
     "@endo/promise-kit": "^1.1.13",
-    "execa": "^9.6.0"
+    "execa": "^9.6.0",
+    "viem": "^2.31.0"
   },
   "devDependencies": {
     "@aglocal/boot": "workspace:*",
@@ -60,8 +61,7 @@
     "c8": "^10.1.3",
     "esbuild": "^0.25.2",
     "ts-blank-space": "^0.6.2",
-    "tsd": "^0.33.0",
-    "viem": "^2.31.0"
+    "tsd": "^0.33.0"
   },
   "ava": {
     "extensions": {
